Separated python and terraform code
