+++ infrastructure/environments/dev/b_configuration/nginx/photos.conf
