+++ terraform/environments/dev/backend.tf
@@ -1,6 +1,6 @@
 terraform {
   backend "gcs" {
-    credentials = "../credentials.json"
+    credentials = "../../../credentials.json"
     prefix  = "01-base"
   }
 }
