+++ terraform/environments/dev/photos.tf
@@ -1,5 +1,5 @@
 module "photos_storage" {
-  source = "../../../terraform/applications/photos/storage"
+  source = "../../applications/photos/storage"
 
   environment = var.environment
   region      = var.region
