+++ terraform/environments/dev/providers.tf
@@ -1,5 +1,5 @@
 provider "google" {
-  credentials = file("../credentials.json")
+  credentials = "../../../credentials.json"
   project     = var.project_id
   region      = var.region
   version     = "~> 3.4"
