+++ terraform/environments/dev/templates/dev_main_init_script.sh
